requirements-api.txt
# FastAPI and API Dependencies
# ============================

# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation and Serialization
pydantic==2.5.0
email-validator==2.1.0.post1

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # For PostgreSQL async support

# WebSocket support
websockets==12.0

# HTTP client for external API calls
httpx==0.25.2
aiohttp==3.9.1

# Background task processing
celery==5.3.4
redis==5.0.1

# Monitoring and Logging
structlog==23.2.0
python-json-logger==2.0.7

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# CORS and Middleware (CORS is built into FastAPI)

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI
pytest-mock==3.12.0

# Optional: Production deployment
gunicorn==21.2.0
